+++ CHANGELOG.md
@@ -5,6 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2025-03-16
+
+| **Build**         | **Status** | **Date**               | **Commit**                              |
+|-------------------|------------|------------------------|-----------------------------------------|
+| P16032025P035000R | Alpha      | 16 March 2025 01:00 PM | updated configuration files and imports |
+
+### Changed
+
+- **package.json**: Removed deprecated `exports` and `jest` configuration for `ConfigSettings` and updated import to
+  `@ConfigSettings`.
+- **playwright.config.ts**: Updated the `baseURL` for tests to use `envConf.configs[envConf.env].frontend.url` instead
+  of hardcoding `https://www.google.com`.
+- **settings/ConfigSettings.ts**: Fixed color logic in `consoleLogs` method for severity levels to ensure proper color
+  coding for `verbose`, `info`, `warning`, and `error` logs.
+- **tsconfig.json**:
+    - Updated TypeScript `Version` to `0.0.2`.
+    - Added new paths for environment configuration and utility functions (e.g., `@envConf`, `@ApiHelper`,
+      `@PlaywrightActions`).
+    - Updated `include` paths to include new directories (`configs`, `src`, `utilities`).
+    - Added `outDir` and `baseUrl` for better module resolution.
+
+### Removed
+
+- **configs/.gitkeep**: Deleted unused file.
+
+### Fixed
+
+- **01-test_steps.ts**: Fixed test step (removed the placeholder `await page.goto("https://www.google.com/");`).
+
 ## [0.0.2] - 2025-03-15
 
 | **Build**         | **Status** | **Date**               | **Commit**                   |
@@ -22,17 +51,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Modified
 
-- `package.json`: Updated project dependencies or configurations.
-- `playwright.config.ts`: Adjusted Playwright configuration.
+- **package.json**: Updated project dependencies or configurations.
+- **playwright.config.ts**: Adjusted Playwright configuration.
 
 ### Added
 
 - `settings/`: Added a new directory for configuration or other settings.
 - `src/frontend/features/01-test.feature`: New feature file for Playwright/Cucumber tests.
 - `src/frontend/step_definitions/01-test_steps.ts`: New step definition file for Playwright/Cucumber tests.
-- `tsconfig.json`: Added a TypeScript configuration file.
+- **tsconfig.json**: Added a TypeScript configuration file.
 
-## [0.0.1]
+## [0.0.1] - 2025-03-14
 
 | **Build**         | **Status** | **Date**               | **Commit**                                                |
 |-------------------|------------|------------------------|-----------------------------------------------------------|